+++ Ink Canvas/Windows/SettingsViews/Pages/AboutPage.xaml
@@ -1,17 +1,17 @@
 <ui:Page
-    x:Class="Ink_Canvas.Windows.SettingsViews2.Pages.FontsPage"
+    x:Class="Ink_Canvas.Windows.SettingsViews.Pages.AboutPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
     xmlns:ikw="http://schemas.inkore.net/lib/ui/wpf"
-    Title="字体设置"
+    Title="关于"
     mc:Ignorable="d">
     <Grid>
         <StackPanel Margin="24">
-            <TextBlock Text="字体设置" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
-            <TextBlock Text="这里是字体设置页面" Style="{DynamicResource BodyTextBlockStyle}" />
+            <TextBlock Text="关于 InkCanvasForClass" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
+            <TextBlock Text="这里是关于页面" Style="{DynamicResource BodyTextBlockStyle}" />
         </StackPanel>
     </Grid>
 </ui:Page>
