+++ Ink Canvas/Windows/SettingsViews/Pages/AboutPage.xaml.cs
@@ -1,6 +1,6 @@
 using iNKORE.UI.WPF.Modern.Controls;
 
-namespace Ink_Canvas.Windows.SettingsViews2.Pages
+namespace Ink_Canvas.Windows.SettingsViews.Pages
 {
     public partial class AboutPage : Page
     {
