+++ Ink Canvas/Windows/SettingsViews/Pages/FontsPage.xaml
@@ -1,17 +1,17 @@
 <ui:Page
-    x:Class="Ink_Canvas.Windows.SettingsViews2.Pages.SettingsPage"
+    x:Class="Ink_Canvas.Windows.SettingsViews.Pages.FontsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
     xmlns:ikw="http://schemas.inkore.net/lib/ui/wpf"
-    Title="设置"
+    Title="字体设置"
     mc:Ignorable="d">
     <Grid>
         <StackPanel Margin="24">
-            <TextBlock Text="设置" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
-            <TextBlock Text="这里是设置页面" Style="{DynamicResource BodyTextBlockStyle}" />
+            <TextBlock Text="字体设置" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
+            <TextBlock Text="这里是字体设置页面" Style="{DynamicResource BodyTextBlockStyle}" />
         </StackPanel>
     </Grid>
 </ui:Page>
