+++ Ink Canvas/Windows/SettingsViews/Pages/HomePage.xaml.cs
@@ -13,7 +13,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace Ink_Canvas.Windows.SettingsViews2.Pages
+namespace Ink_Canvas.Windows.SettingsViews.Pages
 {
     /// <summary>
     /// HomePage.xaml 的交互逻辑
@@ -27,8 +27,8 @@ namespace Ink_Canvas.Windows.SettingsViews2.Pages
 
         private void SettingsCard_Basic_Click(object sender, RoutedEventArgs e)
         {
-            // 找到SettingsWindow2窗口
-            SettingsWindow2 settingsWindow = Window.GetWindow(this) as SettingsWindow2;
+            // 找到SettingsWindow窗口
+            SettingsWindow settingsWindow = Window.GetWindow(this) as SettingsWindow;
             if (settingsWindow != null)
             {
                 // 调用NavigateToPage方法导航到基本设置页面
@@ -38,8 +38,8 @@ namespace Ink_Canvas.Windows.SettingsViews2.Pages
 
         private void SettingsCard_Page2_Click(object sender, RoutedEventArgs e)
         {
-            // 找到SettingsWindow2窗口
-            SettingsWindow2 settingsWindow = Window.GetWindow(this) as SettingsWindow2;
+            // 找到SettingsWindow窗口
+            SettingsWindow settingsWindow = Window.GetWindow(this) as SettingsWindow;
             if (settingsWindow != null)
             {
                 // 调用NavigateToPage方法导航到页面2
@@ -49,8 +49,8 @@ namespace Ink_Canvas.Windows.SettingsViews2.Pages
 
         private void SettingsCard_Design_Click(object sender, RoutedEventArgs e)
         {
-            // 找到SettingsWindow2窗口
-            SettingsWindow2 settingsWindow = Window.GetWindow(this) as SettingsWindow2;
+            // 找到SettingsWindow窗口
+            SettingsWindow settingsWindow = Window.GetWindow(this) as SettingsWindow;
             if (settingsWindow != null)
             {
                 // 调用NavigateToPage方法导航到设计设置页面
@@ -60,8 +60,8 @@ namespace Ink_Canvas.Windows.SettingsViews2.Pages
 
         private void SettingsCard_Appearance_Click(object sender, RoutedEventArgs e)
         {
-            // 找到SettingsWindow2窗口
-            SettingsWindow2 settingsWindow = Window.GetWindow(this) as SettingsWindow2;
+            // 找到SettingsWindow窗口
+            SettingsWindow settingsWindow = Window.GetWindow(this) as SettingsWindow;
             if (settingsWindow != null)
             {
                 // 调用NavigateToPage方法导航到外观设置页面
