+++ Ink Canvas/Windows/SettingsViews/Pages/IconographyPage.xaml.cs
@@ -1,4 +1,4 @@
-namespace Ink_Canvas.Windows.SettingsViews2.Pages
+namespace Ink_Canvas.Windows.SettingsViews.Pages
 {
     public partial class IconographyPage : iNKORE.UI.WPF.Modern.Controls.Page
     {
