+++ Ink Canvas/Windows/SettingsViews/Pages/Page2Page.xaml
@@ -1,9 +1,9 @@
-<Page x:Class="Ink_Canvas.Windows.SettingsViews2.Pages.Page2Page"
+<Page x:Class="Ink_Canvas.Windows.SettingsViews.Pages.Page2Page"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:Ink_Canvas.Windows.SettingsViews2.Pages"
+      xmlns:local="clr-namespace:Ink_Canvas.Windows.SettingsViews.Pages"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
       Title="Page2">
