+++ Ink Canvas/Windows/SettingsViews/Pages/SettingsPage.xaml
@@ -1,17 +1,17 @@
 <ui:Page
-    x:Class="Ink_Canvas.Windows.SettingsViews2.Pages.ThemePage"
+    x:Class="Ink_Canvas.Windows.SettingsViews.Pages.SettingsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
     xmlns:ikw="http://schemas.inkore.net/lib/ui/wpf"
-    Title="主题设置"
+    Title="设置"
     mc:Ignorable="d">
     <Grid>
         <StackPanel Margin="24">
-            <TextBlock Text="主题设置" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
-            <TextBlock Text="这里是主题设置页面" Style="{DynamicResource BodyTextBlockStyle}" />
+            <TextBlock Text="设置" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
+            <TextBlock Text="这里是设置页面" Style="{DynamicResource BodyTextBlockStyle}" />
         </StackPanel>
     </Grid>
 </ui:Page>
