+++ Ink Canvas/Windows/SettingsViews/Pages/ThemePage.xaml
@@ -1,17 +1,17 @@
 <ui:Page
-    x:Class="Ink_Canvas.Windows.SettingsViews2.Pages.AboutPage"
+    x:Class="Ink_Canvas.Windows.SettingsViews.Pages.ThemePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:ui="http://schemas.inkore.net/lib/ui/wpf/modern"
     xmlns:ikw="http://schemas.inkore.net/lib/ui/wpf"
-    Title="关于"
+    Title="主题设置"
     mc:Ignorable="d">
     <Grid>
         <StackPanel Margin="24">
-            <TextBlock Text="关于 InkCanvasForClass" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
-            <TextBlock Text="这里是关于页面" Style="{DynamicResource BodyTextBlockStyle}" />
+            <TextBlock Text="主题设置" Style="{DynamicResource TitleTextBlockStyle}" Margin="0,0,0,16" />
+            <TextBlock Text="这里是主题设置页面" Style="{DynamicResource BodyTextBlockStyle}" />
         </StackPanel>
     </Grid>
 </ui:Page>
