+++ Ink Canvas/Windows/SettingsViews/Pages/TypographyPage.xaml
@@ -1,5 +1,5 @@
 <ui:Page
-    x:Class="Ink_Canvas.Windows.SettingsViews2.Pages.TypographyPage"
+    x:Class="Ink_Canvas.Windows.SettingsViews.Pages.TypographyPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
