+++ Ink Canvas/Windows/SettingsViews/SettingsWindow.xaml
@@ -1,5 +1,5 @@
 <Window
-    x:Class="Ink_Canvas.Windows.SettingsViews2.SettingsWindow2"
+    x:Class="Ink_Canvas.Windows.SettingsViews.SettingsWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
